// frontend/src/api/api8082/models/FeedbackResponseDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authorization server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FeedbackResponseDTO
 */
export interface FeedbackResponseDTO {
    /**
     * 
     * @type {number}
     * @memberof FeedbackResponseDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof FeedbackResponseDTO
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedbackResponseDTO
     */
    category?: string;
    /**
     * 
     * @type {string}
     * @memberof FeedbackResponseDTO
     */
    satisfaction?: string;
    /**
     * 
     * @type {boolean}
     * @memberof FeedbackResponseDTO
     */
    subscribe?: boolean;
    /**
     * 
     * @type {string}
     * @memberof FeedbackResponseDTO
     */
    message?: string;
}

/**
 * Check if a given object implements the FeedbackResponseDTO interface.
 */
export function instanceOfFeedbackResponseDTO(value: object): value is FeedbackResponseDTO {
    return true;
}

export function FeedbackResponseDTOFromJSON(json: any): FeedbackResponseDTO {
    return FeedbackResponseDTOFromJSONTyped(json, false);
}

export function FeedbackResponseDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): FeedbackResponseDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'category': json['category'] == null ? undefined : json['category'],
        'satisfaction': json['satisfaction'] == null ? undefined : json['satisfaction'],
        'subscribe': json['subscribe'] == null ? undefined : json['subscribe'],
        'message': json['message'] == null ? undefined : json['message'],
    };
}

export function FeedbackResponseDTOToJSON(json: any): FeedbackResponseDTO {
    return FeedbackResponseDTOToJSONTyped(json, false);
}

export function FeedbackResponseDTOToJSONTyped(value?: FeedbackResponseDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'category': value['category'],
        'satisfaction': value['satisfaction'],
        'subscribe': value['subscribe'],
        'message': value['message'],
    };
}